+++ src/jcds/transform/__init__.py
@@ -14,7 +14,7 @@ from .convert import (
     to_object,
 )
 
-from .clean import clean_column_names, rename_column, delete_columns, drop_row
+from .clean import clean_column_names, standardize_column_names, rename_column, drop_columns, delete_columns, drop_row
 
 __all__ = [
     "to_int",
@@ -25,9 +25,11 @@ __all__ = [
     "to_bool",
     "to_datetime",
     "to_object",
+    "standardize_column_names",
     "clean_column_names",
     "rename_column",
     "delete_columns",
+    "drop_columns",
     "drop_row",
     "help",
 ]